(BSR)[API] feat: Do no raise error if DMS application is draft
When we receive a webhook notification from DMS about a change on an
application, we validate the application and populate the annotation
if there is an error. But when the application does not have the
annotation field (which is likely for old annotations), the error is
raised and pollutes Sentry.

There is no point in raising the error when the application is draft:
there is nothing we can do. We'll just wait for the application to be
accepted or rejected.