(BSR)[API] ci: Check backend (API) tests before deploying on testing
Before this commit, on a push event on the "master" branch, we always
built and deployed the backend (API), whether or not the "api" folder
was changed, and whether or not the backend tests were successful.

Now the "deploy-to-testing" step requires backend tests to be
successful (or skipped if the backend code did not change).

Also, we now deploy the backend only if it has changed. Ditto for the
frontend.

Finally, we do not need the "build-api" step anymore, since
"tests-api" already builds a Docker image.