(PC-24392)[API] fix: collective offer status filter
Warning: this is a temporary and we-need-it-but-its-not-great fix.

Before: the accepted status values have been slightly changed. This
caused some errors because the filter broke without reason

After: use the previously used allowed values.

Note: * the filtering functions uses a status as a string instead of an
enum which seems odd. This might (should?) be improved.
      * allowing multiple statuses is not great. this should be
        temporary

Co-authored-by: Antoine Marinier <antoine.marinier@passculture.app>