(PC-24954)[EAC] fix: collective api: handle unknown national program
Bug: if a collective offer is updated with an unknown national program
id, then the SQL UPDATE operation will fail. And the error won't be
handled instead of returning a 400 error with a usable error message.

Note: this fix is done quickly. Our session handling comes with a
lot of problems. For example, this specific bug (unknown national
program id) is raised when the update function searches for an
educational institution: a flush is triggered and an error is raised.

We should have clear validation steps done before the real update and
the updated itself should be done using an explicit transaction. But we
need this fix quickly so...