(BSR)[API] fix: Include side effects of booking failure in `book_offer()`
The route called `edit_stock()` upon failure with external tickets,
but changes were not committed. The route could perform the commit,
but it seems more future-proof to call `edit_stock()` (and commit) in
`book_offer()` itself, because calling `edit_stock()` should be the
responsability of `book_offer()`, not its callers.