(PC-25742)[API] fix: Automatically clean up Prometheus metrics directory
When a pod stays alive for a long time, the Prometheus metrics
directory gets populated with a lot of files that take too much
space. It causes the pods to crash. Here we delete metrics file when a
worker is recycled by Gunicorn.

It's possible that Prometheus scrapper does not have enough time to
fetch the latest data before the file is deleted when a worker is
recycled, but we have decided that we can live with that.