(PC-26899)[API] feat: DSv5 in old journey context should garantee the same behavior
In the old journey, if a new DSv4 was created about a venue that was already linked
to a bankInformation, the old bankInformation was updated to the new application details
(especially application ID, IBAN and BIC).
With DSv5 available and both journey active, we should garantee the same behavior.
Hence those modifications, if a new DSv5 is created and we have only one venue
the link those bankInformation, and one already exists, we updating it.