(BSR)[API] feat: Require transaction for destructive statements in pgcli
Destructive statements (any `insert`, `update`, `delete`, `alter
table`, etc.) are now rejected by pgcli if a transaction has not been
started yet (with `begin`). This forces users to start a transaction,
thus allowing a rollback.

Example:

    > alter table mediation validate constraint "mediation_authorId_fkey";
    Destructive statements must be run within a transaction.
    cancelled query

    > begin;
    > alter table mediation validate constraint "mediation_authorId_fkey";
    You're about to run a destructive command.
    Do you want to proceed? [y/N]: y
    Your call!
    ALTER TABLE
    > commit;

Note that pgcli warns the user when they try to exit pgcli while a
transaction is ongoing:

    > begin;
    > quit;
    A transaction is ongoing. Choose `c` to COMMIT, `r` to ROLLBACK, `a` to abort exit. [a]:

This is a standard feature, it is not related to the enabled option.