(BSR)[API] feat: Run `poetry check` when building Docker image
This should detect inconsistencies between `pyproject.toml` and
`poetry.lock`, and will thus refuse to build the Docker image if an
unwanted changed has been added to `pyproject.toml` (without
corresponding changes to `poetry.lock`).

This is mostly useful for local builds, since this check is already
done by a GitHub workflow when building the image for testing and
other environments.