(PC-27760)[API] feat: Add a new test case being in a edge between the old and the new journey
Add a new test for a edge case where an application is created before the new journey is released
but not available for being automatically convert into a bankAccount by the PC-23402 script.

At the time the new bankAccount journey is being released, we'll still have DSv4 applications
in `DRAFT` state. Meaning, users are being filling their application and all informations are not yet
available to import and create a bankAccount (specifically IBAN & BIC, which are mandatory)

In that case, we need to be sure that at the time those applications are set `ON_GOING` (application filled
and being instructed by the compliance) and the new journey is released, a bankAccount is successfully created.