(PC-28448)[EAC] fix: tmplate.contactForm.default: 'form' -> null
Before: default db (and SQLA) value for
collective_offer_template.contactForm was 'form' which lead to some
problems: SQLA model with a null value lead to 'form' saved into
database... which was not what was expected and was against a table
validation check (contactUrl and contactForm cannot both be null.

Why? Because of previous rows. It was decided that the default behaviour
was: the pass culture form has been chosen.

After: default value is null. If the value needs to be 'form', use this
value explicitly.