(PC-28655)[EAC] feat: sync only recent updates
Only ask recently updated venues to Adage API to avoid long and heavy
api requests.

This implies that the deactivation steps must change. Before, we looked
for venues that were not found inside the Adage API response. Now, we
should be a little bit more careful and precise and only target venues
from the current update.