(PC-30109)[API] fix: do not alter venue provider during serialization
Altering a database model during serialization to use Pydantic from_orm
is dangerous because the alterations can be persisted in database during
a autoflush or autocommit.

Those alterations can break database constraints or corrupt further
attempts to serialize the model using from_orm.

As for the price, we must keep reading it from the price rule. Replacing
it by the provider price must be done in a future deployment(#12359).