(PC-30423)[API] chore: Suppress warnings about `Model.query.get` deprecated usage
Custom methods `get()` and `get_or_404()` are defined in the PcObject mixin.
However, some of our models doesn't inherit from it and therefor doesn't benefit
from thoses. Make them inherit from it so we won't have any warnings about
`query.get` deprecated usage.

Side note: This isn't a viable solution for some "dummy" models such as many to many
ones that have a PK made of FKs, i.e. CashflowPricing that
have a PK made of (cashflowId, pricingId). We can live with the fact that for those models
we may have to define `query_class = BaseQuery` in the model, if needed.