(PC-31514)[API] fix: ensure we don't override model's field while serializing
The following code alters the model's field:

    venue.dmsToken = DMS_TOKEN_PRO_PREFIX + venue.dmsToken

This results in an unnecessary UPDATE statement and possibly could affect the
`dmsToken` if it was commited.
Luckily the transaction was rollbacked by default but it is risky and
also creates some deadlocks in the DB.
Since this is mostly for a read only operation, let's remove that
dangerous operation by using Pydantic's getter which returns
basic objects instead of playing with the model's instance.